% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_model_metrics}
\alias{.compute_model_metrics}
\title{Compute Model Metrics}
\usage{
.compute_model_metrics(model_fit, mc_samples, cores, seed)
}
\arguments{
\item{model_fit}{A fitted model object from which posterior samples can be extracted.}

\item{mc_samples}{Integer. The number of Monte Carlo samples to use for approximating the
marginal log-likelihood.}

\item{cores}{Integer. The number of CPU cores to use for parallel computation.}

\item{seed}{Integer. A seed for reproducibility of the Monte Carlo sampling.}
}
\value{
A list containing the following metrics:
\itemize{
\item \code{appx}: The approximate marginal log-likelihood.
\item \code{bic}: The Bayesian Information Criterion (BIC).
\item \code{waic}: The Widely Applicable Information Criterion (WAIC).
}
}
\description{
This function computes various model metrics, including the approximate marginal log-likelihood,
Bayesian Information Criterion (BIC), and Widely Applicable Information Criterion (WAIC),
for a fitted model.
}
\examples{
\dontrun{
# Example usage:
model_fit = ... # A fitted model object
metrics = .compute_model_metrics(
  model_fit = model_fit,
  mc_samples = 1000,
  cores = 4,
  seed = 123
)
print(metrics)
}

}
