% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_models_saved}
\alias{.get_models_saved}
\title{Get Saved Models List}
\usage{
.get_models_saved(save_gdrive, local_path = NULL, gdrive_folder_id = NULL)
}
\arguments{
\item{save_gdrive}{Logical. If \code{TRUE}, the function retrieves the list of models from Google Drive.
If \code{FALSE}, it retrieves the list from a local file.}

\item{local_path}{Character. The local directory path where the models list is stored. Required if
\code{save_gdrive = FALSE}.}

\item{gdrive_folder_id}{Character. The Google Drive folder ID where the models are stored. Required if
\code{save_gdrive = TRUE}.}
}
\value{
A character vector containing the names of the saved models.
}
\description{
This function retrieves a list of saved models either from a local file or from a Google Drive folder.
If the models are stored locally, it reads from a specified text file. If stored on Google Drive,
it lists the files in a specified folder.
}
\examples{
\dontrun{
# Example 1: Retrieve models from a local file
local_path = "/path/to/local/folder"
models_saved = .get_models_saved(
  save_gdrive = FALSE,
  local_path = local_path
)
print(models_saved)

# Example 2: Retrieve models from Google Drive
gdrive_folder_id = "your-google-drive-folder-id"
models_saved = .get_models_saved(
  save_gdrive = TRUE,
  gdrive_folder_id = gdrive_folder_id
)
print(models_saved)
}

}
