% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_save_btblv_models.R
\name{fit_save_btblv_models}
\alias{fit_save_btblv_models}
\title{Fit and Save Bayesian Models for BTBLV}
\usage{
fit_save_btblv_models(
  K_max,
  cluster_run,
  job_cores = NULL,
  job_email = NULL,
  job_name = NULL,
  btblv_data_path,
  iter,
  warmup,
  thin,
  chains,
  precision,
  seed,
  mc_samples,
  config_path,
  model_name_pattern,
  save_gdrive,
  gdrive_folder_id,
  local_path,
  max_treedepth = 10
)
}
\arguments{
\item{K_max}{Integer. The maximum number of clusters to consider.}

\item{cluster_run}{Logical. If \code{TRUE}, the script is executed on a cluster using \code{sbatch}.}

\item{job_cores}{Integer. Number of cores to allocate for the job. Only applicable for cluster runs.}

\item{job_email}{Character. Email address for job notifications. Only applicable for cluster runs.}

\item{job_name}{Character. Name of the job. Only applicable for cluster runs.}

\item{btblv_data_path}{Character. Path to the BTBLV data file.}

\item{iter}{Integer. Number of iterations for model fitting.}

\item{warmup}{Integer. Number of warmup iterations for model fitting.}

\item{thin}{Integer. Thinning interval for model fitting.}

\item{chains}{Integer. Number of chains for model fitting.}

\item{precision}{Character. Precision model type, it can be \code{single} or \code{specific}.}

\item{seed}{Integer. Seed for random number generation.}

\item{mc_samples}{Integer. Number of Monte Carlo samples to generate.}

\item{config_path}{Character. Path to the configuration file.}

\item{model_name_pattern}{Character. Pattern for naming the models.}

\item{save_gdrive}{Logical. If \code{TRUE}, results are saved to Google Drive.}

\item{gdrive_folder_id}{Character. Folder ID for saving results in Google Drive. Required if \code{save_gdrive} is \code{TRUE}.}

\item{local_path}{Character. Local directory path to save results and scripts.}

\item{max_treedepth}{Integer. Maximum tree depth for Stan's Hamiltonian Monte Carlo sampler. Default is 10.}
}
\value{
Character vector. The output from running the bash script.
}
\description{
This function prepares and executes a bash script to fit Bayesian models
for BTBLV data either locally or on a computing cluster. The script is
generated dynamically based on user-specified parameters.
}
\details{
\itemize{
\item When \code{cluster_run} is \code{FALSE}, the script is executed locally using the \code{system} command.
\item When \code{cluster_run} is \code{TRUE}, the script is saved as a bash file and submitted to a cluster via \code{sbatch}.
\item Ensure the folder specified in \code{local_path} exists before running the function.
}
}
\examples{
\dontrun{
  fit_save_btblv_models(
    K_max = 5,
    cluster_run = TRUE,
    job_cores = 4,
    job_email = "user@example.com",
    job_name = "btblv_model",
    btblv_data_path = "data/btblv_data.csv",
    iter = 2000,
    warmup = 1000,
    thin = 1,
    chains = 4,
    precision = 0.01,
    seed = 1234,
    mc_samples = 100,
    config_path = "config.yaml",
    model_name_pattern = "model_K",
    save_gdrive = TRUE,
    gdrive_folder_id = "1A2B3C4D5E6F7G8H9I",
    local_path = "output",
    max_treedepth = 12
  )
}

}
