% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mx_btblv_data}
\alias{mx_btblv_data}
\title{Central mortality rate data}
\format{
An object of class \code{btblv_data} of length 5.
}
\source{
Human Mortality Database.
}
\usage{
mx_btblv_data
}
\description{
\code{btblv::btblv_data} object.
}
\keyword{datasets}
