% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fit_terminal.R
\name{run_fit}
\alias{run_fit}
\title{Run btblv model for different K locally or in a cluster.}
\usage{
run_fit(
  K_max,
  cluster_run,
  job_cores = NULL,
  job_email = NULL,
  job_name = NULL,
  btblv_data_path,
  iter,
  warmup,
  thin,
  chains,
  precision,
  seed,
  mc_samples,
  config_path,
  model_name_pattern,
  save_gdrive,
  gdrive_folder_id,
  local_path
)
}
\arguments{
\item{K_max}{integer with max K to be fitted.}

\item{cluster_run}{logical, if \code{TRUE} fits the model in parallel in the terminal.}

\item{job_cores}{integer with the number of cores.}

\item{job_email}{character with the email for cluster messages.}

\item{job_name}{character with the job name.}

\item{btblv_data_path}{character with the path to the \code{btblv::btblv_data} object.}

\item{iter}{integer with the number of HMC iterations.}

\item{warmup}{integer with the number of warmup iterations.}

\item{thin}{integer with the number of thinning.}

\item{chains}{integer with the number of chains.}

\item{precision}{character with the type of precision. It can be "single" or "specific".}

\item{seed}{integer with the random seed.}

\item{mc_samples}{integer with the number of Monte Carlo samples for the BIC.}

\item{config_path}{character with the config.yaml file path.}

\item{model_name_pattern}{character with the pattern to save the model.}

\item{save_gdrive}{logical indicating if the model should be saved on Google drive.}

\item{gdrive_folder_id}{character with the Google Drive folder id.}

\item{local_path}{character with local folder to save temporary and permanent files.}
}
\value{
\code{TRUE} if the terminal code ran, \code{FALSE} otherwise.
}
\description{
Run btblv model for different K locally or in a cluster.
}
\examples{
 ##
}
