% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fit_btblv.R
\name{save_fit_btblv}
\alias{save_fit_btblv}
\title{Fit btblv model and save it locally or in a Google drive folder}
\usage{
save_fit_btblv(
  btblv_data_path,
  K,
  iter,
  warmup,
  thin,
  chains,
  precision,
  seed,
  mc_samples,
  config_path,
  model_name_pattern,
  save_gdrive,
  gdrive_folder_id,
  local_path,
  ...
)
}
\arguments{
\item{btblv_data_path}{string with path to the \code{btblv::btblv_data} object.}

\item{K}{integer with the latent dimension size.}

\item{iter}{integer with the number of iterations}

\item{warmup}{integer with the warm-up size. See \code{rstan::sampling}.}

\item{thin}{integer with the thinning size. See \code{rstan::sampling}.}

\item{chains}{integer with the number of chains. See \code{rstan::sampling}.}

\item{precision}{string with the type of precision. It is "single" if all
items have the same precision or "specific" if each item has its own precision
parameter. Default is "single".}

\item{seed}{integer with the random seed.}

\item{mc_samples}{integer with the Monte Carlo samples for the marginal likelihood approximation.}

\item{config_path}{string with the path to the config.yaml with credentials for google drive.
Default is \code{NULL}.}

\item{model_name_pattern}{string value with characters to be added to the model name.
It will be used as the name of the final file to be saved.}

\item{save_gdrive}{logical value. If \code{TRUE} saves the model on Google drive.
If \code{FALSE} saves the model locally.}

\item{gdrive_folder_id}{string with the Google drive folder id.}

\item{local_path}{string with the path of the folder used to store local data.
It also needed for the Google drive saving.}

\item{...}{additional \code{rstan::sampling} parameters.}
}
\value{
\code{TRUE} if the model was saved, \code{FALSE} otherwise.
}
\description{
Fit btblv model and save it locally or in a Google drive folder
}
\examples{
##

}
